LayoutTests/svg/dom/svglist-exception-on-out-bounds-error-expected.txt

Tests that out of bounds accesses of SVGList correctly throw exceptions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.getItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.insertItemBefore(null, i) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS svgList.insertItemBefore(seg, i) is seg
PASS svgList.replaceItem(seg, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.replaceItem(null, i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgList.removeItem(i) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS successfullyParsed is true

TEST COMPLETE
